Change font_size to float

+++ config.def.h
@@ -225,7 +225,7 @@ static const bool crop_overscan = true;
 #ifdef HAVE_RMENU
 static const float font_size = 1.0f;
 #else
-static const unsigned font_size = 48;
+static const float font_size = 48;
 #endif
 // Attempt to scale the font size.
 // The scale factor will be window_size / desktop_size.
